requirements.txt
selene==2.0.0rc9
pytest==8.4.1
requests==2.32.4
allure-pytest==2.15.0
dotenv==0.9.9